uint 64

+++ ir-gen/src/node/expression.cc
@@ -15,7 +15,7 @@ QualifiedValue size_of_type( const mty::Qualified *type, Json::Value &ast )
 	return QualifiedValue(
 	  value_type,
 	  Constant::getIntegerValue(
-		value_type->type, APInt( value_type->as<mty::Integer>()->bits, bits / 8, false ) ) );
+		value_type->type, APInt( value_type->as<mty::Integer>()->bits, uint64_t( bits / 8 ), false ) ) );
 }
 
 static QualifiedValue handle_binary_expr( Json::Value &node, VoidType const &_ )
